+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+* Add `--quiet` to `download-file-by-name` and `download-file-by-id`
 
 ### Infrastructure
 * Fix gathering licenses of typeshed libraries